+++ CHANGELOG.md
@@ -141,7 +141,7 @@
 ### New Features
 
 - Add file editing variants of text edit tools plus regex/line-range filtering ([#156](https://github.com/AnswerDotAI/dialoghelper/issues/156))
-- Add `msg_pyrun` tool and `set_pyrun` helper; support async in `_msg_edit` ([#155](https://github.com/AnswerDotAI/dialoghelper/issues/155))
+- Add `msg_python` tool and `set_python` helper; support async in `_msg_edit` ([#155](https://github.com/AnswerDotAI/dialoghelper/issues/155))
 
 
 ## 0.2.13